Issue-11 Thanks to michael.hart.au
In both the NAnt and MSBuild tasks, the logger is set after the Migrator constructor is called, during which the Migrations are loaded - so any logging of the migration loading will not go to the correct logger.
Fixes this by modifying the Migrator constructor to accept an ILogger, or use the current default if none is provided.

git-svn-id: http://migratordotnet.googlecode.com/svn/branches/refactorings-unstable@55 8c3eb3c4-eb3a-0410-862c-73fa8ce6028b